midi popover refactor
