+++ apps/web/src/components/midi-control/types.ts
@@ -1,0 +1,2 @@
+export type DevicePort = { id: string; input: boolean; output: boolean };
+export type Device = { key: string; name: string | null; ports: DevicePort[] };
